dump: Move Plugin struct into its own module
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>
